Create JavaScriptFunction.js
This is my JavaScript code sample that I wrote with an intention of embedding it into my HTML code, on the second file of this repository, which is then used in a form to submit data to a database table, where the data is validated on the client side.

The mechanism of the form used to submit data to a database table is outlined in the next files of this repository. The form contains a range of fields for entering data. This data needs to be validated, on the client side, using JavaScript. When the form is submitted to the server, a page containing PHP scripts extracts the data from the posted information and stores it, using SQL ‘insert’, into the database table.